+++ src/plugins/a11y/index.ts
@@ -1,1 +1,2 @@
 export { a11y } from "./plugin";
+export type { A11yPluginConfig } from "./plugin";