+++ src/plugins/autosize/index.ts
@@ -3,3 +3,4 @@
  */
 
 export { autosize, type AutosizePluginConfig } from "./plugin";
+export type { AutosizeMethods } from "./plugin";